requirements.txt
streamlit>=1.37.0
cerebras-cloud-sdk>=1.0.0
requests>=2.31.0
